kermisattractie af, lukt alleen niet om spin zowel gok als risicovol te maken. voor andere attracties werkt het wel, zo lang ze niet beide zijn

+++ Belastingadviseur.java
@@ -2,9 +2,9 @@ package weekOpdrachtKermis;
 
 public class Belastingadviseur {
 	
-	public static double belastingInnen(double k, double gereserveerdBedrag) {
+	public static int belastingInnen(int k, int gereserveerdBedrag) {
 		System.out.println("De kassa van de kermis heeft voor de belastingen €" + k + " geld in de kassa.");
-		double KermisKassa = k - gereserveerdBedrag;
+		int KermisKassa = k - gereserveerdBedrag;
 		System.out.println("De belasting die betaalt moet worden is " + gereserveerdBedrag);
 		System.out.println("De kassa van de kermis heeft nu €" + KermisKassa + " geld in de kassa.");
 		return KermisKassa;
