+++ Ladderklimmen.java
@@ -1,20 +1,20 @@
 package weekOpdrachtKermis;
 
 public class Ladderklimmen extends Attractie implements GokAttractie{
-	private double gereserveerdBedrag =0;
+	private int gereserveerdBedrag =0;
 	
-	Ladderklimmen(String naam, double prijs) {
+	Ladderklimmen(String naam, int prijs) {
 		super(naam, prijs);
 	}
 	
-	public double kansSpelBelastingBetalen(double kassa){
+	public int kansSpelBelastingBetalen(int kassa){
 		gereserveerdBedrag += (this.getPrijs()*0.3);
 //		System.out.println("kassa is "+ (this.getPrijs() - (this.getPrijs()*0.3))); // check of berekening klopt
 		kassa += (this.getPrijs() - (this.getPrijs()*0.3));
 		return kassa;
 	}
 	
-	public double getGereserveerdBedrag() {
+	public int getGereserveerdBedrag() {
 		return gereserveerdBedrag;
 	}
