+++ Spookhuis.java
@@ -2,7 +2,7 @@ package weekOpdrachtKermis;
 
 public class Spookhuis extends Attractie{
 
-	Spookhuis(String naam, double prijs) {
+	Spookhuis(String naam, int prijs) {
 		super(naam, prijs);
 	}
